+++ cmem_plugin_reason/utils.py
@@ -11,10 +11,10 @@ from xml.etree.ElementTree import Element, SubElement, tostring
 from cmem.cmempy.dp.proxy.graph import get_graph_import_tree, post_streamed
 from cmem.cmempy.dp.proxy.sparql import post as post_select
 from cmem.cmempy.dp.proxy.update import post as post_update
+from cmem_plugin_base.dataintegration.context import ExecutionContext
 from cmem_plugin_base.dataintegration.description import PluginParameter
-from cmem_plugin_base.dataintegration.parameter.choice import ChoiceParameterType
 from cmem_plugin_base.dataintegration.parameter.graph import GraphParameterType
-from cmem_plugin_base.dataintegration.plugins import ExecutionContext, WorkflowPlugin
+from cmem_plugin_base.dataintegration.plugins import WorkflowPlugin
 from cmem_plugin_base.dataintegration.types import BoolParameterType, IntParameterType
 from defusedxml import minidom
 
@@ -36,24 +36,16 @@ MAX_RAM_PERCENTAGE_DEFAULT = 20
 ONTOLOGY_GRAPH_IRI_PARAMETER = PluginParameter(
     param_type=GraphParameterType(classes=["http://www.w3.org/2002/07/owl#Ontology"]),
     name="ontology_graph_iri",
-    label="Ontology_graph_IRI",
+    label="Ontology graph IRI",
     description="The IRI of the input ontology graph.",
-)
-
-REASONER_PARAMETER = PluginParameter(
-    param_type=ChoiceParameterType(REASONERS),
-    name="reasoner",
-    label="Reasoner",
-    description="Reasoner option.",
-    default_value="elk",
 )
 
 MAX_RAM_PERCENTAGE_PARAMETER = PluginParameter(
     param_type=IntParameterType(),
     name="max_ram_percentage",
     label="Maximum RAM Percentage",
-    description="""Maximum heap size for the Java virtual machine in the DI container running the
-    reasoning process. ⚠️ Setting the percentage too high may result in an out of memory error.""",
+    description="""Maximum heap size for the reasoning process in the DI container. ⚠️ Setting the
+    percentage too high may result in an out of memory error.""",
     default_value=MAX_RAM_PERCENTAGE_DEFAULT,
     advanced=True,
 )
@@ -65,6 +57,21 @@ VALIDATE_PROFILES_PARAMETER = PluginParameter(
     description="""Validate the input ontology against OWL profiles (DL, EL, QL, RL, and Full) and
     annotate the result graph.""",
     default_value=False,
+)
+
+OUTPUT_GRAPH_IRI_PARAMETER = PluginParameter(
+    param_type=GraphParameterType(
+        allow_only_autocompleted_values=False,
+        classes=[
+            "https://vocab.eccenca.com/di/Dataset",
+            "http://rdfs.org/ns/void#Dataset",
+            "http://www.w3.org/2002/07/owl#Ontology",
+        ],
+    ),
+    name="output_graph_iri",
+    label="Output graph IRI",
+    description="""The IRI of the output graph for the inconsistency validation. ⚠️ Existing graphs
+    graphs will be overwritten.""",
 )
 
 
@@ -86,15 +93,15 @@ def create_xml_catalog_file(dir_: str, graphs: dict) -> None:
 
 
 def get_graphs_tree(graph_iris: tuple) -> dict:
-    """Get graph import tree"""
+    """Get graph import tree. Last item in tuple is output_graph_iri which is excluded"""
     graphs = {}
-    for graph_iri in graph_iris:
+    for graph_iri in graph_iris[:-1]:
         if graph_iri not in graphs:
             graphs[graph_iri] = f"{token_hex(8)}.nt"
             tree = get_graph_import_tree(graph_iri)
             for value in tree["tree"].values():
                 for iri in value:
-                    if iri not in graphs:
+                    if iri not in graphs and iri != graph_iri[-1]:
                         graphs[iri] = f"{token_hex(8)}.nt"
     return graphs
 
@@ -113,25 +120,24 @@ def send_result(iri: str, filepath: Path) -> None:
 
 def post_provenance(plugin: WorkflowPlugin, prov: dict | None) -> None:
     """Post provenance"""
-    if not prov:
-        return
-    param_sparql = ""
-    for name, iri in prov["parameters"].items():
-        param_sparql += f'\n<{prov["plugin_iri"]}> <{iri}> "{plugin.__dict__[name]}" .'
-    insert_query = f"""
-        INSERT DATA {{
-            GRAPH <{plugin.output_graph_iri}> {{
-                <{plugin.output_graph_iri}> <http://purl.org/dc/terms/creator>
-                    <{prov["plugin_iri"]}> .
-                <{prov["plugin_iri"]}> a <{prov["plugin_type"]}>,
-                    <https://vocab.eccenca.com/di/CustomTask> .
-                <{prov["plugin_iri"]}> <http://www.w3.org/2000/01/rdf-schema#label>
-                    "{prov['plugin_label']}" .
-                {param_sparql}
+    if prov:
+        param_sparql = ""
+        for name, iri in prov["parameters"].items():
+            param_sparql += f'\n<{prov["plugin_iri"]}> <{iri}> "{plugin.__dict__[name]}" .'
+        insert_query = f"""
+            INSERT DATA {{
+                GRAPH <{plugin.output_graph_iri}> {{
+                    <{plugin.output_graph_iri}> <http://purl.org/dc/terms/creator>
+                        <{prov["plugin_iri"]}> .
+                    <{prov["plugin_iri"]}> a <{prov["plugin_type"]}>,
+                        <https://vocab.eccenca.com/di/CustomTask> .
+                    <{prov["plugin_iri"]}> <http://www.w3.org/2000/01/rdf-schema#label>
+                        "{prov['plugin_label']}" .
+                    {param_sparql}
+                }}
             }}
-        }}
-    """
-    post_update(query=insert_query)
+        """
+        post_update(query=insert_query)
 
 
 def get_provenance(plugin: WorkflowPlugin, context: ExecutionContext) -> dict | None:
@@ -142,14 +148,14 @@ def get_provenance(plugin: WorkflowPlugin, context: ExecutionContext) -> dict | 
     project_graph = f"http://di.eccenca.com/project/{context.task.project_id()}"
 
     type_query = f"""
-            SELECT ?type ?label {{
-                GRAPH <{project_graph}> {{
-                    <{plugin_iri}> a ?type .
-                    <{plugin_iri}> <http://www.w3.org/2000/01/rdf-schema#label> ?label .
-                    FILTER(STRSTARTS(STR(?type), "https://vocab.eccenca.com/di/functions/"))
-                }}
+        SELECT ?type ?label {{
+            GRAPH <{project_graph}> {{
+                <{plugin_iri}> a ?type .
+                <{plugin_iri}> <http://www.w3.org/2000/01/rdf-schema#label> ?label .
+                FILTER(STRSTARTS(STR(?type), "https://vocab.eccenca.com/di/functions/"))
             }}
-        """
+        }}
+    """
 
     result = json.loads(post_select(query=type_query))
 
@@ -198,7 +204,7 @@ def get_provenance(plugin: WorkflowPlugin, context: ExecutionContext) -> dict | 
 def robot(cmd: str, max_ram_percentage: int) -> CompletedProcess:
     """Run robot.jar"""
     jar = Path(__path__[0]) / "bin" / "robot.jar"
-    cmd = f"java -XX:MaxRAMPercentage={max_ram_percentage} -jar {jar} " + cmd
+    cmd = f"java -XX:MaxRAMPercentage={max_ram_percentage} -jar {jar} {cmd}"
     return run(shlex.split(cmd), check=False, capture_output=True)  # noqa: S603